+++ model-classes.lisp
@@ -231,15 +231,13 @@ message-descriptor.")
   (make-instance 'option-descriptor
                  :name name :value value :type type))
 
-;; todo(jgodbout): Is t needed?
 (defun find-option (desc name)
   "Given a protobuf descriptor DESC and the NAME of an option, returns the
-   value of the option and its Lisp type. The third value is T if an option was
-   found, otherwise NIL."
+   value of the option and its Lisp type, otherwise NIL."
   (declare (type descriptor desc) (type string name))
   (let ((option (find name (proto-options desc) :key #'proto-name :test #'option-name=)))
     (when option
-      (values (proto-value option) (proto-type option) t))))
+      (values (proto-value option) (proto-type option)))))
 
 (defgeneric remove-options (descriptor &rest names)
   (:documentation